- Added working frontend routes - Made tables CSS code dry - Cleaned up files and folders - Updating Navbar links using new routes
