+++ frontend/src/App.css
@@ -11,6 +11,6 @@ body {
 }
 
 .all-tables {
-  margin-top: 310px;
+  margin-top: 270px;
   margin-bottom: 100px;
 }
