+++ frontend/src/components/LeadsTable.js
@@ -1,23 +1,20 @@
 import React, { Component } from 'react';
-import './tablestyles/globalTableStyles.css';
-import './tablestyles/leaderboard.css';
-import Navbar from '../Navbar'
-import FiscalYear from '../FiscalYear';
+import '../styles/globalTableStyles.css';
+import FiscalYear from './FiscalYear';
 
-class EmployeeLeaderboard extends Component {
+class LeadsTable extends Component {
     render() {
         return (
             <div>
-                <Navbar />
-                
-                <div class="leaderboard_table">
-                    <h1 class="header_leaderboard header">EMPLOYEE LEADERBOARD</h1>
-                    <div class="CMP_leaderboard">
-                        <h3>CaptureMortgage+ Leaderboard by Employee</h3><span> </span>
+
+                <div class="center">
+                    <h1 class="header leads">LEADS</h1>
+                    <div class="leads">
+                        <h3>CaptureMortgage+ Leads Board</h3><span> </span>
                         <h3><FiscalYear /></h3>
                     </div>
                     <table id="myTable">
-                        <tr class="table_header_leaderboard">
+                        <tr class="leads-back">
                             <th>ID</th>
                             <th>Entered as Lead</th>
                             <th>Customer Name</th>
@@ -28,38 +25,39 @@ class EmployeeLeaderboard extends Component {
                         </tr>
                         <tr>
                             <td>1</td>
-                            <td>01/03/2019</td>
-                            <td>Alicia Keys</td>
-                            <td>Re-finance</td>
-                            <td>$400,000</td>
+                            <td>01/01/2019</td>
+                            <td>Craig David</td>
+                            <td>Home Loan</td>
+                            <td>$200,000</td>
                             <td>LP Staff</td>
                             <td>Sarah Smith</td>
                         </tr>
                         <tr>
                             <td>2</td>
-                            <td>01/02/2019</td>
-                            <td>Beyonce</td>
+                            <td>01/01/2019</td>
+                            <td>James Dean</td>
                             <td>Home Loan</td>
-                            <td>$260,000</td>
+                            <td>$40,000</td>
                             <td>LP Staff</td>
                             <td>Louise Lilly</td>
                         </tr>
                         <tr>
                             <td>3</td>
-                            <td>01/01/2019</td>
-                            <td>Craig David</td>
-                            <td>Re-finance</td>
-                            <td>$110,000</td>
+                            <td>02/04/2018</td>
+                            <td>Billy Joel</td>
+                            <td>Home Loan</td>
+                            <td>$475,000</td>
                             <td>LP Staff</td>
                             <td>Katherine Kale</td>
                         </tr>
                     </table>
                 </div>
             </div>
+                
         );
     }
 }
 
-export default EmployeeLeaderboard;
+export default LeadsTable;
