+++ frontend/src/styles/crm.css
