--- frontend/src/tables/tablestyles/approvalsTable.css
@@ -1,15 +1,0 @@
-.CMP_approvals {
-    text-align:center;
-    color: #F19D38;
-  }
-
-.header_approvals {
-    text-align:center;
-    color: #F19D38;
-
-  }
-
-.table_header_approvals {
-    background:#F19D38;
-    color: white;
-}
