--- frontend/src/tables/tablestyles/leaderboard.css
@@ -1,14 +1,0 @@
-.CMP_leaderboard {
-    text-align:center;
-    color: #0066CC;
-  }
-
-.header_leaderboard {
-    text-align:center;
-    color: #0066CC;
-}
-
-.table_header_leaderboard {
-    background:#0066CC;
-    color: white;
-}
