--- frontend/src/tables/tablestyles/leads.css
@@ -1,14 +1,0 @@
-.CMP_leads {
-    text-align:center;
-    color: #0066CC;
-  }
-
-.header_leads {
-    text-align:center;
-    color: #0066CC;
-}
-
-.table_header_leads {
-    background:#0066CC;
-    color: white;
-}
