--- frontend/src/tables/tablestyles/lodgements.css
@@ -1,14 +1,0 @@
-.CMP_lodgements {
-    text-align:center;
-    color: #66CCCC;
-  }
-
-.header_lodgements {
-    text-align:center;
-    color: #66CCCC;
-  }
-
-.table_header_lodgements {
-    background:#66CCCC;
-    color: white;
-}
