--- frontend/src/tables/tablestyles/settlements.css
@@ -1,14 +1,0 @@
-.CMP_settlements {
-    text-align:center;
-    color: #43963F;
-  }
-
-.header_settlements {
-    text-align:center;
-    color: #43963F;
-  }
-
-.table_header_settlements {
-    background:#43963F;
-    color: white;
-}
